requirements.txt
jira==3.0.1
Flask~=2.0.1
flask-mongoengine==1.0.0
celery~=5.1.2
matplotlib==3.4.3
mongoengine==0.23.1
slack-sdk==3.11.1
gunicorn==20.1.0
redis==3.5.3
python-dotenv==0.19.0


